Basic functionality command arg
Added a -t command arg to stop the shutdown calls on respective systems but print the time in hours, minutes, seconds and then in seconds. This is mainly for users who want the data output but not the shutdown itself. I came across this need when using WSL on windows for a game I was downloading, and the shutdown would just close the WSL program rather than my system